vinfo: add join checks to QP mode
Due to probesize issues, the ffprobe parser in the QP mode may be restarted.
This causes the initial few frames to be parsed twice. The join between the
generic ffprobe run and the QP run is very brittle, as it just assumes
consecutive frames from both lists refer to the same frame (and it is OK
with lists with different sizes).

The solution consists of:
* (1) make sure the 2x lists have the same number of elements
* (2) make sure the common fields in both elements are the same (an
  exception is the resolution value, which in the QP case does not
  consider cropping, and in the generic case does)
* (3) understand probesize-based restarts, and reset the QP list when
  that happens.

This patch covers (1) and (2). It also does some refactoring of the QP
list calculation to make (2) easier.